+++ deck.h
@@ -161,11 +161,15 @@ typedef std::map<std::string, long double> DeckList;
 class Decks
 {
 public:
-    void add_deck(Deck* deck, const std::string& deck_name);
-    Deck* find_deck_by_name(const std::string& deck_name);
+    ~Decks();
+
     std::list<Deck> decks;
     std::map<std::pair<DeckType::DeckType, unsigned>, Deck*> by_type_id;
     std::map<std::string, Deck*> by_name;
+
+    void add_deck(Deck* deck, const std::string& deck_name);
+    Deck* find_deck_by_name(const std::string& deck_name);
+
 };
 
 #endif