+++ sim.h
@@ -183,7 +183,7 @@ struct CardStatus
     const Card* m_card;
     unsigned m_index;
     unsigned m_action_index;
-    Field* m_field;
+    Field* m_field; // only needed for fixes/updated skills in attack power calculation
     unsigned m_player;
     unsigned m_delay;
     unsigned m_hp;